Backend/requirements.txt
fastapi
uvicorn
motor
beanie
python-dotenv
pyjwt
bcrypt
pydantic
passlib
email-validator
orjson
ujson
httptools
watchfiles
aiostream
requests
bs4
redis 
aiohttp
pymongo
